Clean up emacs backups. Gitinclude empty dir
Git only tracks file. So to include an empty directory you must
create a nonce file inside the directory and track that file.
You can modify your .gitignore to ignore everything else inside the
directory.

This is useful for directories like ~/.emacs.d/backups because on
a new installation the absence of the directory may cause errors
while emacs tries to save backups. But you don't want the files
that you edit to be visible to the world because they are being
tracked by git.