hledger: Different major modes for editing/viewing
The same keymap is shared among all the buffers
in a major-mode. I had bound "q" to kill buffer in buffers
used to view hledger-mode reports. The buffers were in hledger-mode
so that the account names are colored similarly. But this changed
the keymap of the buffers being used for editing the file and hence
made entering "q" quite a challenge. So, I create a mode that is
a complete copy of hledger-mode but doesn't share the same keymap
object.