A simple setup for reading in emacs.
Turn on evil-local-mode in the buffer for navigation.  I just like the
keybindings for navigation that we have in vim. What I don't like about
the navigation is the lack of preserving the cursor position as and when
we scroll using Ctrl+D or Ctrl+U. Maybe I am more used to the Emacs way
of doing things and hence it seems more intuitive to me.