Show expanded hledger running report
Alongside the concise report about the top-level expenses and income
source, it is useful to have the detailed report so that you can see the
distribution of money among the sub accounts. It would have been great
if I could implement a widget-like interface where a click could expand
a node into its sub categories but I am willing to commit the required
time for it.