Fix coding issues with dictionary lookup.
It seems that after you decode the contents of a buffer and you have the
decoded string (say in utf-8), inserting it again into the buffer
doesn't ensure that you will see decoded text in the buffer. I am not
sure why this is the case but that is what I see. The good thing is that
can parse json directly from a string as well. Coding systems are a hard
nut to crack.