Fix unicode encoding issues with Mailgun client.
Encode every multipart/form-data section value with
(decode-coding-string value 'utf-8). If not encoded,
emacs seems to insert the unibyte/multibyte representation
of the string value which confuses the server for some reason
I don't know. It confuses the server to the extent that it cannot
parse any of the section, not even the ones that precede the faulty
section.