Switch to htmlize.el from htmlfontify.el
Gmail is said to strip off all the contents of the head in an HTML
email. So, we need to have inline css. htmlize.el lets you do this very
easily. The problem now is that the text is still line-wrapped on my
mobile phone. It causes the multi-column reports to be messed up.