Setup to send monthly reports by email.
I am using async to run a separate emacs instance which loads up all
of my configuration. Since I do this once every month, I think it's
okay. If the process fails, I re-try every 5 minutes. This can be tricky
if we have the hledger journal file in an incosistent state while we are
trying to prepare the reports. So, let's see when that happens. :)