Fix hledger mail async task
It is not good to load everything up in the async task because the
current active interactive instance of emacs may be using the exact same
files and this might cause the task to fail while the rest of the
mailing functionality worked. This would have causes a thousands of
emails to be sent to the user.