Remove region or local journal related functions.
It is usually not as useful to be able to perform operations on a local
journal. In case you are managing more than one journal files, this
would be important. Currently, I don't know of a straightforward way to
implement this. The whole codebase requires a bit of refactoring.

I would like to do the following things:

1. Use processes to get ther results of running the hledger binary
   instead of letting the process modify the buffer itself.
2. Making the overall report a bit more customizable so that somebody
   can leave a few things out that he doesn't find usefuall at all.