Fix quote visiblity in scratch buffer.
If the old quote before quote refresh was visible in the *scratch*
buffer, it seems reasonable to expect that the new quote should also be
visible. This might move the visible position of the point in the buffer
a bit by re-centering it inside the window. It's okay because the
refresh happens after a long idle time.