Facility for balace percentages.
Since hledger doesn't have the feature to display the percentages of the
balances in a report, I wrote a function that would look around point to
find number and show their percentage relative to their sum. This won't
work if the balances are displaed in a tree like fashion as the parent
accounts balance would be added twice and would mess up the percentages.