Fix summary for emergency fund ratio.
Emergency fund ratio is not the number of months one would sustain
with the current lifestyle but with a lifestyle where he only spends
on the critically necessary. Introducing a new fund ratio, total fund
ratio, which is the number of months that one's liquid assets would
last if continues living the same lifestyle.