Remove paredit in ruby-mode. Setup initial window config.
`paredit-mode' isn't good for major-modes where you don't want a
leading space before pairs that paredit inserts, except for the pairs
inserted in the beginning of a line. So, paredit-mode makes sure that
before every '(' there is some whitespace. This isn't good for
programming languages where in you would like to call functions with
syntax `f(arg1)' and do not like the style of having a space between
`f' and the `('. I like to reserve this space for constructs like the
`for' and `while'. I developed this habit when I started coding in C
back in college.