Refactor scratch buffer refresh to use text properties.
The previous definition used a global variable to keep track of the
size of the quote text. Hence if the quote text was deleted and some
important text entered before a refresh happens, that important text
would be deleted. Using text properties, only quote text that has
their ~:text-type~ property set to ~:quote-string~ are is deleted.